\name{DPSPN-package}
\alias{DPSPN-package}
\alias{DPSPN}
\docType{package}
\title{
\packageTitle{DPSPN}
}
\description{
\packageDescription{DPSPN}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DPSPN}
\packageIndices{DPSPN}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{DPSPN}

Maintainer: \packageMaintainer{DPSPN}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
